Splitting the IssueListener into 2 different "features"
