+++ app/config/github.yml
@@ -1,7 +1,17 @@
 # listener services that you can opt into for a repository
 services:
-    app.subscriber.issue_subscriber:
-        class: AppBundle\Issues\IssueListener
+    app.subscriber.status_change_by_comment_subscriber:
+        class: AppBundle\Subscriber\StatusChangeByCommentSubscriber
+        arguments:
+            - '@app.status_api'
+
+    app.subscriber.needs_review_new_pr_subscriber:
+        class: AppBundle\Subscriber\NeedsReviewNewPRSubscriber
+        arguments:
+            - '@app.status_api'
+
+    app.subscriber.bug_label_new_issue_subscriber:
+        class: AppBundle\Subscriber\BugLabelNewIssueSubscriber
         arguments:
             - '@app.status_api'
 
@@ -10,8 +20,12 @@ parameters:
     repositories:
         symfony/symfony:
             subscribers:
-                - app.subscriber.issue_subscriber
+                - app.subscriber.status_change_by_comment_subscriber
+                - app.subscriber.needs_review_new_pr_subscriber
+                - app.subscriber.bug_label_new_issue_subscriber
             # secret: change_me
         symfony/symfony-docs:
             subscribers:
-                - app.subscriber.issue_subscriber
+                - app.subscriber.status_change_by_comment_subscriber
+                - app.subscriber.needs_review_new_pr_subscriber
+                - app.subscriber.bug_label_new_issue_subscriber
