+++ src/AppBundle/Subscriber/NeedsReviewNewPRSubscriber.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace AppBundle\Subscriber;
+
+use AppBundle\Event\GitHubEvent;
+use AppBundle\GitHubEvents;
+use AppBundle\Issues\Status;
+use AppBundle\Issues\StatusApi;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+
+class NeedsReviewNewPRSubscriber implements EventSubscriberInterface
+{
+    private $statusApi;
+
+    public function __construct(StatusApi $statusApi)
+    {
+        $this->statusApi = $statusApi;
+    }
+
+    /**
+     * Adds a "Needs Review" label to new PRs.
+     *
+     * @param GitHubEvent $event
+     */
+    public function onPullRequest(GitHubEvent $event)
+    {
+        $data = $event->getData();
+        $repository = $event->getRepository();
+        if ('opened' !== $action = $data['action']) {
+            $event->setResponseData(array('unsupported_action' => $action));
+
+            return;
+        }
+
+        $pullRequestNumber = $data['pull_request']['number'];
+        $newStatus = Status::NEEDS_REVIEW;
+
+        $this->statusApi->setIssueStatus($pullRequestNumber, $newStatus, $repository);
+
+        $event->setResponseData(array(
+            'pull_request' => $pullRequestNumber,
+            'status_change' => $newStatus,
+        ));
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return array(
+            GitHubEvents::PULL_REQUEST => 'onPullRequest',
+        );
+    }
+}
